pyproject.toml
[project]
name = "stream-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cupy-cuda12x>=13.6.0",
    "ipython>=9.7.0",
    "numcodecs>=0.16.3",
    "numpy>=2.3.4",
    "nvidia-nvcomp-cu12>=5.0.0.6",
    "nvtx>=0.2.13",
    "rmm-cu12>=25.10.0",
    "zarr>=3.1.3",
]
